famtree_dj/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
# cffi==1.15.1
# charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==40.0.2
defusedxml==0.7.1
dj-database-url==1.3.0
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
django-rest-passwordreset==1.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
# PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
